Implement master data service in docker-compose, update README to reflect new service status, and refactor models by moving customer, salesperson, process, size, and color definitions to master_data.go. Adjust response metadata to use int64 for total count.

+++ internal/models/response.go
@@ -12,10 +12,10 @@ type Response struct {
 
 // Meta 分页元数据
 type Meta struct {
-	Page       int `json:"page"`
-	PageSize   int `json:"page_size"`
-	Total      int `json:"total"`
-	TotalPages int `json:"total_pages"`
+	Page       int   `json:"page"`
+	PageSize   int   `json:"page_size"`
+	Total      int64 `json:"total"`
+	TotalPages int   `json:"total_pages"`
 }
 
 // ErrorResponse 错误响应
